requirements.txt
annotated-types          0.7.0
anyio                    4.6.2.post1
attrs                    24.2.0
beautifulsoup4           4.12.3
cachetools               5.5.0
certifi                  2024.8.30
cffi                     1.17.1
charset-normalizer       3.4.0
colorama                 0.4.6
distro                   1.9.0
google                   3.0.0
google-api-core          2.21.0
google-api-python-client 2.149.0
google-auth              2.35.0
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.1
google-cloud             0.34.0
googleapis-common-protos 1.65.0
h11                      0.14.0
httpcore                 1.0.6
httplib2                 0.22.0
httpx                    0.27.2
idna                     3.10
jiter                    0.6.1
oauth2client             4.1.3
oauthlib                 3.2.2
openai                   1.52.0
outcome                  1.3.0.post0
pip                      24.2
proto-plus               1.24.0
protobuf                 5.28.2
pyasn1                   0.6.1
pyasn1_modules           0.4.1
pycparser                2.22
pydantic                 2.9.2
pydantic_core            2.23.4
pyparsing                3.2.0
PySocks                  1.7.1
python-dotenv            1.0.1
requests                 2.32.3
requests-oauthlib        2.0.0
rsa                      4.9
selenium                 4.25.0
selenium-stealth         1.0.6
setuptools               75.2.0
six                      1.16.0
sniffio                  1.3.1
sortedcontainers         2.4.0
soupsieve                2.6
tqdm                     4.66.5
trio                     0.27.0
trio-websocket           0.11.1
typing_extensions        4.12.2
undetected-chromedriver  3.5.5
uritemplate              4.1.1
urllib3                  2.2.3
websocket-client         1.8.0
websockets               13.1
wsproto                  1.2.0